% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{plot.sumcat}
\alias{plot.sumcat}
\title{Plot}
\usage{
\method{plot}{sumcat}(x, y, ...)
}
\arguments{
\item{x}{A sumcat list object which is the summary of classifications
made by four different Classification methods. Can be reproduced with the
model_cat() function.}

\item{y}{A null value in this case as all the information is stored inside
the x object which is a sumcat object.}
}
\value{
A density plot made with functions provided by the ggplot2 package
}
\description{
Takes the fitted values(actual values) less the
  predicted values creating the residuals. These are then plotted as a
  density curve over one another, using the geom_density from ggplot2.
}
\details{
Plot function which visualize residual plots from several methods for
  classification. Returns a plot with each residual density plot colored.
}
\examples{
 x <- model_cat(Potability~., water_potability)
 plot(x)

}
