% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rank Function.R
\name{rank}
\alias{rank}
\title{rank}
\usage{
rank(formula, train, test, response)
}
\arguments{
\item{formula}{The formula that the user wants to use for prediction.}

\item{train}{The training dataset used for prediction.}

\item{test}{The test dataset used for prediction.}

\item{response}{The response variable from the the test dataset. This is the
variable we are trying to predict.}
}
\description{
Ranks several methods of classification based on their level of accuracy.
Returns a matrix of methods and ranked accuracies, with higher accuracies
listed first.
}
