% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rank Function.R
\name{rank.sumcat}
\alias{rank.sumcat}
\title{rank}
\usage{
rank.sumcat(object)
}
\arguments{
\item{object}{An object passed from the "model_cat" function.}
}
\description{
Ranks several methods of classification based on their level of accuracy.
Returns a matrix of methods and ranked accuracies, with higher accuracies
listed first.
}
