% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_function.R
\name{rank_methods}
\alias{rank_methods}
\title{rank_methods}
\usage{
rank_methods(object)
}
\arguments{
\item{object}{An object passed from the "model_cat" function.}
}
\description{
Ranks several methods of classification based on their level of predictive
accuracy. Returns a matrix of methods and ranked accuracies from largest to
smallest. MUST run "model_cat" function prior to "rank_methods".
}
\examples{
object <- model_cat(Potability ~ ., water_potability, water_test)
ranking(object)
}
