% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_function.R
\name{ranking}
\alias{ranking}
\title{rank}
\usage{
ranking(object)
}
\arguments{
\item{object}{An object passed from the "model_cat" function.}
}
\description{
Ranks several methods of classification based on their level of accuracy.
Returns a matrix of methods and ranked accuracies, with higher accuracies
listed first.
}
\examples{
object <- model_cat(Potability ~ ., water_potability, water_test)
ranking(object)
}
