% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_sumcat.R
\name{summary.sumcat}
\alias{summary.sumcat}
\title{summary.sumcat}
\usage{
\method{summary}{sumcat}(sumcat)
}
\arguments{
\item{sumcat}{This is a sumcat object which is a list of the different
classification models embedded inside.}
}
\description{
Displays all the models one after the other, specifying the call that was
  made by displaying those results to the user in the console.
}
\examples{
  x <- model_cat(Potability ~ ., water_potability, water_test)
  summary(x)
}
