% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{water_test}
\alias{water_test}
\title{Potability of Water Test Set}
\format{
A data frame with 403 rows and 10 variables
}
\source{
\url{https://www.kaggle.com/adityakadiwal/water-potability}
}
\usage{
data(water_test)
}
\description{
A dataset containg the potability of water among other 9 other
 parameters of more than 400 observations.
}
\details{
\describe{
  \item{ph}{pH of water (0 to 14)}
  \item{Hardness}{Capacity of water to precipitate soap in mg/L}
  \item{Solids}{Total dissolved solids in ppm}
  \item{Chloramines}{Amount of Chloramines in ppm}
  \item{Sulfate}{Amount of Sulfates dissolved in mg/L}
  \item{Conductivity}{Electrical conductivity of water
    in microsiemens per centimeter}
  \item{Organic_carbon}{Amount of organic carbon in ppm}
  \item{Trihalomethanes}{Amount of Trihalomethanes in micrograms per liter}
  \item{Turbidity}{Measure of light emitting property of water
    in NTU}
  \item{Potability}{Indicates if water is safe for human
    consumption. Potable -1 and Not potable -0}
}
}
\keyword{datasets}
