% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_cat.R
\name{model_cat}
\alias{model_cat}
\title{Modeling function for Ensemble of Categorical Analysis}
\usage{
model_cat(formula, data, test_data = water_test)
}
\arguments{
\item{formula}{A formula statement}

\item{data}{A categorical data set specified by the user}
}
\value{
A list object that contains the ensemble of models
}
\description{
Creates a 'sumcat' class object that when summary is called the models
  for each type: Random Forest, Logistic Regression, LDA, SVM.
}
