% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{plot.sumcat}
\alias{plot.sumcat}
\title{Plot}
\usage{
\method{plot}{sumcat}(x)
}
\arguments{
\item{x}{Variable with is dependent from the other variables:
"Variable name ~."}

\item{train}{A dataset, which is used to train the models of several
classification methods}

\item{test}{A dataset, which is used to test the models in order to
create predictions}
}
\description{
Plotfunction which visualize residualplots from several methods for
classification. Returns a different colored plot
}
