% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.sumcat.R
\name{rank.sumcat}
\alias{rank.sumcat}
\title{rank.sumcat}
\usage{
rank.sumcat(object)
}
\arguments{
\item{object}{An object passed from the "model_cat" function.}
}
\description{
Ranks several methods of classification based on their level of predictive
accuracy. Returns a matrix of methods and ranked accuracies from largest to
smallest. MUST run "model_cat" function prior to "rank.sumcat".
}
